refactor: centralize formatBytes/formatElapsed into utils/format.ts
Eliminates 7 copies of identical byte/elapsed formatting functions
across 6 TUI files, replacing them with a single ANSI-free pure
module at src/utils/format.ts. Also updates PRD-0012 status to
[DONE] and adds the missing websearch multi-provider changeset.

Resolves: H3 and M4 from architecture improvement report.
